postgres/pyproject.toml
[project]
name = "postgres"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "databricks-sdk>=0.62.0",
]
